fix(node): update stale block time after resetting connections
If a block is considered stale, the `advance_state` function will reset
connections with all peers. To reset the "staleness" of a block, the
`LastBlockMonitor::update` function must be called. Currently that is done
upon receiving new block headers, but no headers will ever be sent, as
`advance_state` disconnects the peers before they can send them once
a block is considered stale. Here we patch this to reset the "staleness"
of a block after disconnecting from all peers.