build: new msrv job to isolate build tools
The msrv job checks just that the library code can be built with the
MSRV's compiler, and that is it. All other build tooling functionallity
doesn't need to be gated on the MSRV since it doesn't effect consumers.
We specifically would rather use cargo's V3 resolver by default to
generate our lockfiles.