build: run basic feature matrix in CI
Extending the "build" step to run all feature flags, as well as
the existing default. Could possible add more flag combos in the future
if things get complicated, but this should cover most cases for now.

Switched to the optional dependency syntax which ensures that the
(unused in our case) implicit features of the optional dependencies are
not created. This cleans up the library interface for consumers, only i
our explicit flags are exposed.

Dropping the "check" CI step since build covers all of it.