build: run basic feature matrix in CI
Extending the "build" step to run all feature flags, as well as
the existing default. Could possible add more flag combos in the future
if things get complicated, but this should cover most cases for now.
This helps ensure that the different flag combos exposed by the library
actually build for consumers. If the feature flag matrix begins to get
more complex, we could look into a more specific tool like cargo-hack,
but for now simplicity is fine.

Switched to the optional dependency syntax which ensures that the
(unused in our case) implicit features of the optional dependencies are
not created. This cleans up the library interface for consumers, only i
our explicit flags are exposed.

Dropping the "check" CI step since build covers all of it. If build gets
really slow in the future, we can add check back as a form of "fast
failure", but as it is today this is just duplicating work and slowing
CI down.