refactor: move `LastBlockMonitor` and `NodeState` out of `node` module
The `node` module is getting cluttered and these are self-contained structures.
Doing a bit of a stretch in encapsulation here, but I think since the time we see
the last block is implicit to the node network connections, I would like it to be
in `network`. The `NodeState` is currently very primitive, so I think `lib` is suitable.
If `NodeState` starts to contain data or methods then it might belong in `node` again.