fix(network): introduce Socks5 proxy connection
The Arti project does not have an official MSRV, and the dependency list has
a number of failed security audits caught by the CI job in this repository.
The legacy Tor client is sufficient for desktop and some mobile users, and
no developers have expressed interest in using Arti directly. A simple Socks5
proxy that connects to a Tor daemon allows for Tor connections without an
explicit dependency that causes MSRV conflicts, security audits, and lib-sqlite
version conflicts.