build: move build complexity to justfile
Keep build and test logic as agnostic as possible to whatever CI task
runner is used (today that is Github Actions). Let the CI job focus on
performance, like caching artifacts and keep the build and test logic in
the forefront easy for developers to use.

With that goal in mind, the CI job has been refactored to a handful
of jobs which each call a `just` recipe. Consistent caching has been
added to each job which shows a nice improvement in speed for the
workflow, running in less than a minute. The justfile has been
refactored with the logic that previously lived in the CI yaml. This
makes it easier to maintain as well as run locally when developing. A
few un-used recipes have been cleaned up.

Lockfile complexity has been relegated to the _test-msrv recipe. This
recipe checks that there is at least one set of dependencies which
satisfy the msrv requirements. The check no longer depends on the
checked in lockfile which doesn't matter much since consumers create
their own.

A rust-toolchain.toml file was added as a gently nudge for contributers
to stay on a recent version of "stable", that way we can use some of the
newer features like the v3 resolver and see less of "works on my
machine". CI will most likely not use the file however, since it usually
needs to test more complex things like MSRV.