fix(client)!: use unbounded channels to avoid dropped requests
The `UnboundedSender` offers two advantages over bounding the number
of requests a client can make. First it drops the need to maintain an
`async` and sync version of the client methods - `UnboundedSender::send`
is synchronous. This is actually very convient for `ldk-node`, where
the client will have to implement a synchronous trait.

Secondly there are some methods that are critical to succeed if the node
is running. Adding a script must always work if the node is still online.
Because it is unacceptable to miss some messages, there is basically two
approaches: cache the failures and try again later, or use an unbounded
channel. Since caching the failures will result in the exact same overhead
as stored messages in a channel when the node is not ready to accept them
yet, the unbounded channel is much easier to implement and reason about.