feat(justfile): add options to delete data and git branches
When branches are merged into master, most of the time there is
no point in holding on to them locally. There is also a directory
that stores node data when the examples sync. It can be helpful to
remove this and resync from scratch when testing. Also if people
want to switch toolchains they might want to delete the lockfile
and rebuild it.