fix(chain): avoid requesting duplicate blocks
The node may request filters of multiple peers to
speed up the filter syncing process. If a filter is
checked and the desired block is fetched, the current
code will request that block again if the filter is
seen again (from a laggy peer). Instead, only add blocks
to the request queue for the filters we haven't already
 checked