fix(lib): improve `filter-control` API and functionality
First, we make sure that the same filter is not emitted twice,
so as to not waste the client computation. Second, make it possible
to round trip from `IndexedFilter`, to raw bytes, back to `BlockFilter`.
Also remove a pointless structure from `lib.rs`

This was missed when considering duplicate filters in
the usual flow. The same is desirable for the managed filter
workflow, where the same filter should not be emitted twice.