refactor(chain): remove `await` points when syncing CF Headers/Filters
With the new assumptions about `BlockTree`, such that every header will
be available in memory when syncing CFHeaders and CFilters, there is no
need to check the database for heights/hashes that are not present in
the `BlockTree`. This removes the need for `await` points in most
functions. The `send_chain_update` function is called by `node`,
when processing CFHeader and CFilter messages.

This starts towards a cleaner split between sync/async methods and
has seemingly increased performance on my end. Of course the cost
is all the relevant headers must be stored in the `BlockTree` for
a sync, but for most cases, even a recovery of 2/3 years can easily
be held in RAM by most phones.