refactor(chain): get all relevant headers at start
There is logic to re-initialize the `BlockTree` if we need to evaluate
a reorganization, but this header should just be loaded into the tree
to begin with. It is also desirable to compute the difficulty adjustment
of course, so if we have a relevant header for that, it should also be
loaded into memory.