feat: add additional timeout parameter
Potentially the most important timeout of them all,
we move on from connections that can't open a TCP
stream within a timeout. This was hardcoded as two
seconds, but should actually be more like the millisecond
scale in production. This commit adds this initial
timeout as a parameter to `PeerTimeoutConfig`