refactor!: remove `FilterSyncPolicy`
The original intention with this configuration was so a silent-payments
wallet could fetch key tweaks for a block before downloading filters.
I was experimenting with a silent-payments sync, and in practice this
is actually a bad idea. The disk space required to store key tweaks for
a full block would far exceed the size of the filter. This API path
should be discouraged, and instead the implementation should store the
filters however they like, then request the key tweaks and dispose of
them. This also cleans up the syncing state-machine and allows a removal
of a mostly pointless integration test.