fix(network): emit `Info` after a tx gossip
The traditional paradigm of submitting a transaction to a server and
consdering it "sent" doesn't really apply to the node. If a user is
broadcasting a new transaction, it cannot possibly be in another node's
mempool, so they will request a `getdata`. At this point, it is fairly
certain the transaction will propagate, but the transaction may still be
rejected in the future for having insufficient fee or non-standard
struture. The best we can do is say "this transaction was gossiped" but
there is no notion of this transaction is sent and will definitely be in
the remote mempool.