refactor(network): introduce `ReaderMessage`
The `Reader` produces a subset of the `NetworkMessage` that are relevant
to the light client, omitting things like compact block relay. The
condensed message is called `PeerMessage`, which is passed to both
`Peer` and the main thread. The `Peer` is doing some filtering as well.
For instance the `Peer` can respond to `Ping` messages, and the main
thread does not need to know about them. The `PeerMessage` should be
anything that is worthy of going to the main thread, which in the ideal
case is block headers, compact block filters, filters, inventory
messages of block annoucements. Everything else can be handled by
`Peer`, including address gossip at some point.