test: add transaction broadcast test
AFICT the only way to build a transaction that the Regtest instance does
not known about is to send money to a key outside of the Regtest wallet,
then broadcast a transaction signed from that key. Unfortunately that
makes this test pretty nasty in terms of implementation. I think the
only thing I may have overlooked is the RPC `createrawtransaction` but
that RPC still involves some inputs that might be difficult to fetch.
Thankfully there is a `rust-bitcoin` example of signing a transaction,
so I went ahead and yanked a bunch of logic there.

ref: https://github.com/rust-bitcoin/rust-bitcoin/blob/master/bitcoin/examples/sign-tx-taproot.rs