refactor(network): introduce `MessageState`
The state for the messages a `Peer` sends and receives is mostly handled
in `counter.rs`. The API is rudimentary and prone for errors, and was
just used as a simple starter for DoS protection and basic sanity
checks. This begins a refactor towards a more expressive state, where
the API is stricter, and more questions may be asked about the message
state with a given peer. For instance, a simple thing we might want to
know is if the version handshake is complete.

To mock test the timeout related functionality, the `test-utils` module
is brought in from the development dependency on `tokio`, which allows
for instant "sleep" calls.