fix(network)!: strictly handle reject messages
Continual refactor to remove items from `counter.rs`. When we send a
transaction, we only expect a single reject message, if any. The
`MessageState` now closely tracks which transactions were broadcast and
if we have already been informed that they were rejected. This prevents
the remote peer from sending random reject messages.

Unfortunately this is a breaking change in the `RejectPayload`.