refactor(network): add gossip state for `addr` msgs
Another field removed from `counter` to define more explicitly the
state of peer address gossip. In the usual message flow, we send a
`getaddr` message at the beginning of the connection with a peer. This
is normally met with 250-300 potential peers. When this occurs, we move
the state from `NotReceived` to `RandomGossip`. In the random gossip
phase, we expect to only receive 1-5 peers every minute or two. The
number of peers a remote node is allowed to gossip is capped at 10_000,
which is subject to change.