feat(client)!: return accepted/disconnected headers
At the moment the only information returned by `BlocksDisconnected` is
the blocks that were reorganized, but some sources of the data may also
be able to make use of the accepted headers. For instance they may
simply replace the old value at the height with the new one. Here we
return that data, as well as make the ordering explicit in terms of
height.