fix(network): time messages with `TimeSensitiveId`
The previous `counter` design relied on specific methods to be called
for each message type. Instead each time-sensitive message variant can come with its
own unique identifier, and these can be added/removed from the message
state as they are sent/received. Messages for which time is not
relevant simply do not report an ID. This simplifies the logic for
`Peer`, as time sensitive messages have a single unified call when
sending/receiving them.

For the actual implementation, we don't really care about the content of
a `headers` or `cfheaders` message, just that they respond. For blocks,
we can ID the request by the block hash. Filters require some extra
thought, as 1_000 filters are sent for a single filter request. It is
unlcear what exactly the "timeout" should be for the pace of the filter
download