fix(network): add timeout for filter messages
Filter messages typically take longer than the other chain messages, so
they will likely have to have their own timeout in the future. In the
meantime, the remote node should still have a limited time to respond
with at least one filter. This is mostly useful for the "always on" use
case.