feat: use DNS as a source for initial peers
Presumably, some end-users of this crate will download network data
infrequently. This means the node may have to iterate through a high
number of stale peers within the peer store, which ultimately leads to a
poor user experience. The DNS seeders are acting as always-on network
crawlers. If a developer is comfortable with trusting these seeders,
they can provide high quality potential connections. In my opinion this
isn't much of a privacy tradeoff, as the DNS resolver only knows the
user is interested in the bitcoin network. This is, however, a
significant trust tradeoff, as now the DNS seeders are used as a
source-of-truth for good bitcoin nodes.

Guidance on if this feature should be recommended or not depends on some
network analysis. @nyonson if your research finds that the proportion of
filter-serving nodes _listening_ for inbound connections is low enough,
this would be okay to promote to end-developers. On the other hand, if
the network is reasonably available to serve filters, this could be used
in special cases where we don't expect the node to shut down frequently.

This is controversial in general, so this will never be a default.