refactor(network): make module errors ergonomic
There are a million pointless `map_err` around the module that I can
only assume I added out of laziness. This module has grown quiet a bit,
so the errors should be easier to work with. I don't implement `source`
here because these are not used in a backtrace. Instead they are used to
drive control flow.