feat(network): add `PingState`
An occasional ping message should be sent if we haven't received a
message from a connection of ours for a while. It doesn't matter what
message this is, so the time can be updated for every message we
receive. If enough time passes, we generate a random nonce to send to
our peer and wait to receive this nonce back to reset the state. The
actual timing of the `Ping`/`Pong` message will be handled using
`TimeSensitiveMessageId`, however the state ensures we get the correct
nonce echoed back.

This feature only really applies to always-on nodes, and the time until
we send a ping can be generous. Here is is defined as two minutes.