feat: add `get_fees_for_block` fn
Fetching the fee rates for a block involves fetching all `Transaction`
corresponding to a `OutPoint`, accumulating the total value spent in the
transaction, and dividing by the weight once all input values are known.
This function assumes that a channel of `Transactions` will be made to
receive these transactions relevent to the `OutPoint`. To look up what
transaction the `OutPoint` was in after receiving a tranasction, a map of
the outpoint `Txid` corresponding to where they were spent is
maintained. Once it is known what transaction the feerate is being
calculated for, the total input value may be adjusted. Once all input
values are known, the final `FeeRate` may be returned. When all
`FeeRate` are known, the value is propagated to the client.

Because this is as an atomic task that has a well defined start and end,
I adopt a more functional style here, similar to `connect_socks5`