network: set up fetch transaction state
To fetch fee rates the node will need a way to issue a command to start
fetching the fees, some state as to what transactions we requested, and
a way to actually handle the transactions from the `reader`. Assuming
there is some channel to receive the transactions along with their
`Txid`, the `Sender` is pre-factored into the command to start fetching
transactions.