refactor(chain): move out filter message state
I am adopting the idea that the `Node` should drive message state
forward, and the `Chain` should simply report changes and hold state
about the chain. This is already the case for `CFHeaderChanges`, where
`Node` decides to request the next batch of headers, or not, based on
what the result of the `CFHeader` message was.

I am adopting the same attitude for `sync_filter`, which now reports:
1. should this `block_hash` be requested
2. was this the filter corresponding to `stop_hash`

In this way, `Node` decides if the next message should be made, and will
also allow `Node` to manage the `BlockQueue`, which will come next.

It should also be noted that `is_filters_synced` was called for every
filter message, but this iterates over the whole tree. We should only
do this if we are considering to build the next message or not. This
early return condition isn't worth the cost.