feat!: add new chain `Info` messages
The chain height has been neglected in the `Info` messages and may be
useful to end users that are recovering a wallet. Currently there is no
way to know the node is actually doing something during the header sync
without looking at the string logs. I would imagine most apps will not
and should not display the string logs directly. This adds a couple
messages about the header chain that fall into the `Info` category.