refactor: move `BlockQueue` up to `Node`
The `BlockQueue` can be used for more than requesting positive filter
checks. Namely the client may use the block queue to requests blocks to
calculate fee rates. To avoid refactors in the future, instead of moving
the `BlockQueue` to the `PeerMap` or elsewhere, it is moved up to the
`Node` level.