fix(testnet4): do not compute difficulty adjustment
Raised by: https://github.com/bitcoindevkit/bdk-cli/pull/197

Users of `Testnet4` will run into problems when loading headers from the
database, as the network allows for minimum difficulty blocks. Instead
of adding logic for the header timestamps just for the sake of
`Testnet4`, I am opting to simply ignore the difficutly adjustment. This
boolean does not evaluate to `true` on `Bitcoin`, which is easily
verified in `rust-bitcoin`.

Notably `Signet` does not allow minimum difficulty blocks, so difficulty
adjustment logic is still tested by a testnet as well as mainnet. This
is only dropping `Testnet4`