refactor(net): DNS bootstrap at the start of node
When trying to select a random peer from the database, if the database
happens to have 0 peers available, and there are no trusted peers, then
the DNS bootstrap occurs. This special logic does not need to occur when
selecting a peer to connect to. Instead, the node may boostrap if
necessary at the start of the process, given there are no peers
available.