refactor: allow any client to request a block
I have not confirmed, but I suspect some users in the lightning context
would be able to make use of requesting blocks for channel annoucements.
While it cannot be proven that the channel still exists, if a channel
output is known as well as the block hash, this function would prove the
channel is in the best chain. From poking around other wallets, I think
this would also potentially be useful in Liana.