refactor: allow multiple requests per block
For fee rate calculation the `BlockQueue` must account that the user may
be interested in a block twice, once for computing a fee rate and once
because the scripts are interesting to the client. This is also now true
because of the previous commit, which allows any client to request a
block.

Line 462 and 475 ensure we only add the block hash to the request queue
a single time if a filter is deemed interesting, so the
`BlockQueue::add` does not need to check for containment. When receiving
a block off of the wire, the client either:
1. wants the block, so it should be sent to them
2. already got the block and this was a stale request
3. is a valid block but was not requested by the client

The `ProcessBlockResponse` now covers each of these scenarios. By far
the most common will be 1., and it is uncertain if 3. is even possible
because the `Peer` does sanity-checking for blocks requested as well.