lib: add `lookup_host` DNS function
The node census revealed that there is a low proportion of nodes serving
compact block filters, and, further still, a low number of nodes
actually listening for inbound connections. The DNS seeders solve these
problems at the expense of trusting the seeder that it is not serving
bullshit nodes. The user experience of using compact block filters is
still up for question, and users may decide they are comfortable
providing a seed that they trust to make their syncing experience
faster.

Instead of determining within `Node` when or how DNS will be queried,
besides in the initial bootsrap phase, this function can just export
some of the internal functionality already present in the `dns` modules.
Developers and users can decide if/how they would like to use DNS.

This PR removes an ignored test and sanity-checks the `lookup_host`
function within the integration tests.