chain: add block subsidy calculation
Taken directly from Bitcoin Core, this function computes the current
block subsidy given a height in the chain. With the subsidy, the average
fee rate in a block may be computed as `(coinbase outputs(s) - subsidy) / block weight`.

The average fee rate for a block may still prove useful for fee
estimations by the light client, so a simple function is added and
tested here.

ref: https://github.com/bitcoin/bitcoin/blame/master/src/validation.cpp#L1942