feat(client): fetch average feerate for block
Building on the last commit, the total fee rate for a block can be
computed as long as the subsidy is known. A new method is added on the
client which uses the existing fetch block functionality, but adds the
steps to compute the average fee rate for the block once it is actually
fetched. Also the requirements are loosened for removing blocks from the
queue, and the node will still request blocks even if the relevant
filters have been checked.