feat: use BIP-130 `sendheaders`
According to the BIP, nodes may, but are not obligated to, announce new
blocks by sending a `headers` message. From the perspective of the light
client, we don't care about their preference for this, because it is not
relaying blocks. However, the light may use this method as a form of
hearing about new blocks.

No matter the current state of the client, if headers are announced the
node state should no reset to `NodeState::Behind`. Otherwise, the
`BlockTree` and will be able to make sense of the new headers, and the
node may proceed with the subsequent steps of the sync.