lib: Remove user-accessable debug logs
Now that there are some initial users with `bitcoin-safe` and the
BDKSwiftExampleApp, I have seen a few integrations and found some
glaring problems with the API. In practice, the debug
(`LogLevel::Debug`) strings are not seen by the end-user, and developers
are not using them either. When they have been used, I have seen that
developers are parsing these strings for information. That would imply
these strings are "versioned" in a sense. These should be removed ASAP
and only accessable by developers of this repository.

I believe these are still useful in the debugging process during
testing, so I have these strings emitted under `cfg(test)`. A little
hacky, but it serves the purpose as what these debug strings have
typically been used for.