lib: Remove internal header checkpoints
Now that a few use-cases have popped up, wallet implementations have
leaned to the safe side on wallet recovery. My original intention for
the header checkpoints was for users with new wallets. In practice this
is not as helpful for a couple reasons

1. Users will probably have to query a server for fees anyway. If they
   are building a new wallet for a user, they might as well take a more
   recent checkpoint to start the sync. By the time developers adopt the
   latest release, the checkpoints in the binary have gotten old.
2. Any developer supporting hardware wallets has no choice but to scan
   from Segwit or _maybe_ Taproot
3. The checkpoints are only as useful as the frequency of releases,
   which will go down as the repository goes through refactoring

Not to mention the checkpoints are a trusted component of the binary
(personally don't think it's controversial, but still nice to minimize
trust)