lib: Remove `filter-control` and adding scripts
Recovering a wallet, or perhaps more complex use-cases like managing a
LN node, involves uncertainty with regards to what scripts to add to the
node on startup. With the default API, as scripts are found as inclusions
in blocks, they would have to add them to the node as it runs. This is a
race condition, as the next filter(s) may have already been downloaded.

The following algorithm is the only one that is safe to my knowledge
when recovering a wallet.
1. query a filter with `N` existing scripts
2. if a match is found, reveal a next script `j`
3. check the filter for `j`
4. if `j` is in the filter, reveal another script `k` and repeat 2.
   otherwise stop

This is because the wallet at hand may have two or more scripts in a
block, especially in the case of larger wallets or LN routing nodes.
While it would be possible to somehow implement this logic, the better
solution is to let users query these filters as many times as they would
like.

Reducing the scope of resposibility in the node is a net win, and this
was a particularly nasty case that cropped up. By removing the script
responsibility, I want to take the opportunity to clean up other
components as well.