node: Make `Node::run` mutable
This was a strange design choice I made a while ago related to the
language bindings. There is no reason to put these behind shared
resources, as there is only one source-of-truth for all of these fields,
the main loop. My hope is removing all of these locks saves a lot of
CPU cycles. There is also no reason to not just consume the node when
calling `run`. I am reasonable confident this will work with minimal
changes in the bindings.

While I was at it, use the typical `Interval` from `tokio` and increase
the loop speed. Also, to increase redudancy during the filter sync, the
next filter request is broadcasted to all peers. This ensures we
continue to sync even if a node drops off.