lib: Hide `NodeState` as internal
This enum is a bit of API creep as this is really used as an internal
driver. Users are able to request blocks at their own discretion, and
the previous `Synced` event might be emitted even when the user has
blocks they are interested. The most information we can disclose is that
the filters have been synced to the network. Everything else is
emitted via `Info` or `Event`

This allows future models of the state machine to not break existing
users.

I also add an early return to `sync_filter`, as we now broadcast the
filter requests to all peers, so some peers may be sending duplicates