Change crate name to `bip157`
The discoverability of `kyoto-cbf` on crates.io is probably not great.
One would either need to know "Kyoto" is a light client implementation,
or that CBF stands for compact block filters. Even if they know the term
Kyoto, there are some `kyoto` crate naming clashes that might also be
misleading.

What is nice about BIP-157 is that the BIP itself defines the client
protocol. By taking this name, it should at least be implied to the user
what the crate does. Any future release of compact filter related
sub-crates would probably fall under BIP-158.

The goal is to go from Google search to crates.io or docs.rs when a user
is interested in BIP-157/158. While I will keep the repository name, I
think changing the crate might be a net benefit. There is a nice synergy
with the `bip324` naming, which also does P2P stuff, and `rust-bitcoin`
modules like `bip32` and `bip152`.