refactor!: Remove the `HeaderStore` trait
Preparation to complete remove I/O related operations with relation to
block headers. Since the introduction of `BlockTree` all headers are
held in memory and have constant time access. The client can decide what
they would like to do with block headers. There is no need for a trait
in this scenario.