Remove the SQL peer storage 🗑️
In an effort to limit externally maintained dependencies, `rusqlite`
should be removed. This PR removes it and replaces the logic with the
`bitcoin-address-book`. That crate ensures a single source of gossip
cannot fill the entire address book, so this crate can add peers to the
tables with less logic for eclipse resistance.

The tables of the address managers are held in memory, so operations are
much faster than previously. Instead of having the convoluted logic for
spawning internal tasks, peers can add them directly.