Remove the `SqliteHeaderDb` from `Chain`
I found a weird edge case where Bitcoin Core will not respond if you
request filter commitments/filters for the genesis hash. Otherwise this
breaks all tests that attempt to check reorged blocks are caught when
starting on an orphan block. Follow up will be a the event-based system
so the client can save headers.