Introduce `ChainState` to `Builder`
Users are now expected to persist block header information. This allows
them to build a parse or dense chain, and decide on the I/O method they
want to use. As a result, they will need to provide some initial
starting values. Either a checkpoint, likely for a new wallet, or a
recent snapshot.