Remove `rusqlite` dependency
I looked/experimented with BDK, particularly the `bdk_chain`, and I
think [this
method](https://docs.rs/bdk_chain/latest/bdk_chain/struct.CheckPoint.html#method.insert)
is sufficient to remove the store entirely instead of deprecating it.
Since every block header is now emitted, the `bdk_kyoto` integration can
make use of this to avoid starting on an orphan block. Reorganizations
don't need to be explicitly emitted. Rather, the node can just emit
whatever chain it is on and `CheckPoint::insert` will reconsile the
changes.

It looks like there will also be nice [user facing
events](https://github.com/bitcoindevkit/bdk_wallet/pull/310), so
explicitly informing about a re-org might even be returned in the wallet
update.

This Kyoto now acts as a decision-maker for the the chain of most work,
but there is now a single source-of-truth for the history of headers.